# Throw Caution to the Wind or: How I Learned To Stop Worrying And Love The Race