--- a/lib/entry-index.js
+++ b/lib/entry-index.js
@@ -29,8 +29,16 @@ function insert (cache, key, digest, opts) {
       time: +(new Date()),
       metadata: opts.metadata
     }
+    const stringified = JSON.stringify(entry)
+    // NOTE - Cleverness ahoy!
+    //
+    // This works because it's tremendously unlikely for an entry to corrupt
+    // another while still preserving the string length of the JSON in
+    // question. So, we just slap the length in there and verify it on read.
+    //
+    // Thanks to @isaacs for the whiteboarding session that ended up with this.
     return appendFileAsync(
-      bucket, '\n' + JSON.stringify(entry)
+      bucket, `\n${stringified.length}\t${stringified}`
     ).then(() => entry)
   }).then(entry => (
     fixOwner.chownr(bucket, opts.uid, opts.gid).then(() => (
@@ -46,10 +54,16 @@ function find (cache, key) {
   let ret
   return Promise.fromNode(cb => {
     pipe(stream, split('\n', null, {trailing: true}).on('data', function (l) {
+      const pieces = l.split('\t')
+      if (!pieces[1] || pieces[1].length !== parseInt(pieces[0], 10)) {
+        // Length is no good! Corruption ahoy!
+        return
+      }
       let obj
       try {
-        obj = JSON.parse(l)
+        obj = JSON.parse(pieces[1])
       } catch (e) {
+        // Entry is corrupted!
         return
       }
       if (obj && (obj.key === key)) {
@@ -92,9 +106,14 @@ function lsStream (cache) {
                 if (err) { return cb(err) }
                 const entries = {}
                 data.split('\n').slice(1).forEach(function (entry) {
+                  const pieces = entry.split('\t')
+                  if (pieces[1].length !== parseInt(pieces[0], 10)) {
+                    // Length is no good! Corruption ahoy!
+                    return
+                  }
                   let parsed
                   try {
-                    parsed = JSON.parse(entry)
+                    parsed = JSON.parse(pieces[1])
                   } catch (e) {
                   }
                   // NOTE - it's possible for an entry to be