--- a/lib/entry-index.js
+++ b/lib/entry-index.js
@@ -5,7 +5,6 @@ const contentPath = require('./content/path')
 const crypto = require('crypto')
 const fixOwner = require('./util/fix-owner')
 const fs = require('graceful-fs')
-const lockfile = require('lockfile')
 const path = require('path')
 const pipe = require('mississippi').pipe
 const Promise = require('bluebird')
@@ -14,64 +13,30 @@ const through = require('mississippi').through
 
 const indexV = require('../package.json')['cache-version'].index
 
+const appendFileAsync = Promise.promisify(fs.appendFile)
+
 module.exports.insert = insert
 function insert (cache, key, digest, opts) {
   opts = opts || {}
   const bucket = bucketPath(cache, key)
-  const lock = bucket + '.lock'
   return fixOwner.mkdirfix(
     path.dirname(bucket), opts.uid, opts.gid
-  ).then(() => (
-    Promise.fromNode(_cb => {
-      const cb = (err, entry) => {
-        lockfile.unlock(lock, er => {
-          _cb(err || er, entry)
-        })
-      }
-      lockfile.lock(lock, {
-        stale: 60000,
-        retries: 10,
-        wait: 10000
-      }, function (err) {
-        if (err) { return _cb(err) }
-        fs.stat(bucket, function (err, existing) {
-          if (err && err.code !== 'ENOENT' && err.code !== 'EPERM') {
-            return cb(err)
-          }
-          const entry = {
-            key: key,
-            digest: digest,
-            hashAlgorithm: opts.hashAlgorithm,
-            time: +(new Date()),
-            metadata: opts.metadata
-          }
-          // Because of the way these entries work,
-          // the index is safe from fs.appendFile stopping
-          // mid-write so long as newlines are *prepended*
-          //
-          // That is, if a write fails, it will be ignored
-          // by `find`, and the next successful one will be
-          // used.
-          //
-          // This should be -very rare-, since `fs.appendFile`
-          // will often be atomic on most platforms unless
-          // very large metadata has been included, but caches
-          // like this one tend to last a long time. :)
-          // Most corrupted reads are likely to be from attempting
-          // to read the index while it's being written to --
-          // which is safe, but not guaranteed to be atomic.
-          const e = (existing ? '\n' : '') + JSON.stringify(entry)
-          fs.appendFile(bucket, e, function (err) {
-            cb(err, entry)
-          })
-        })
-      })
-    })
-  )).then(entry => {
-    return fixOwner.chownr(bucket, opts.uid, opts.gid).then(() => {
-      return formatEntry(cache, entry)
-    })
-  })
+  ).then(() => {
+    const entry = {
+      key: key,
+      digest: digest,
+      hashAlgorithm: opts.hashAlgorithm,
+      time: +(new Date()),
+      metadata: opts.metadata
+    }
+    return appendFileAsync(
+      bucket, '\n' + JSON.stringify(entry)
+    ).then(() => entry)
+  }).then(entry => (
+    fixOwner.chownr(bucket, opts.uid, opts.gid).then(() => (
+      formatEntry(cache, entry)
+    ))
+  ))
 }
 
 module.exports.find = find
@@ -126,7 +91,7 @@ function lsStream (cache) {
               fs.readFile(path.join(indexDir, bucket, f), 'utf8', function (err, data) {
                 if (err) { return cb(err) }
                 const entries = {}
-                data.split('\n').forEach(function (entry) {
+                data.split('\n').slice(1).forEach(function (entry) {
                   let parsed
                   try {
                     parsed = JSON.parse(entry)
@@ -186,30 +151,15 @@ function bucketDir (cache) {
 module.exports._bucketPath = bucketPath
 function bucketPath (cache, key) {
   const hashed = hashKey(key)
-  return path.join(bucketDir(cache), hashed.slice(0, 2), hashed)
+  return path.join(bucketDir(cache), hashed.slice(0, 2), hashed.slice(2))
 }
 
 module.exports._hashKey = hashKey
 function hashKey (key) {
-  // NOTE (SECURITY)
-  //
-  // `sha1` conflicts can be generated, but it doesn't matter in this case,
-  // since we intend for there to be regular conflicts anyway. You can have
-  // the entire cache in a single bucket and all that'll do is just make a big
-  // file with a lot of contention, if you can even pull it off in the `key`
-  // string. So whatever. `sha1` is faster and it doesn't trigger the warnings
-  // `md5` tends to (yet?...).
-  //
-  // Not to mention, that in the case of pacote/npm, the amount of control
-  // anyone would have over this key is so minimal that it's incredibly
-  // unlikely that they could intentionally generate a large number of
-  // conflicts just with a package key such that they'd do anything resembling
-  // a hash flood DOS.
   return crypto
-  .createHash('sha1')
-  .update(key.toLowerCase()) // lump case-variant keys into same bucket.
+  .createHash('sha256')
+  .update(key)
   .digest('hex')
-  .slice(0, 7)
 }
 
 function formatEntry (cache, entry) {