--- a/test/util/cache-index.js
+++ b/test/util/cache-index.js
@@ -26,7 +26,7 @@ function CacheIndex (entries, hashAlgorithm) {
       if (typeof lines.length !== 'number') {
         lines = [lines]
       }
-      serialised = lines.map(JSON.stringify).join('\n')
+      serialised = '\n' + lines.map(JSON.stringify).join('\n')
     }
     insertContent(tree, parts, serialised)
   })