Fixes: #21 

This is a pretty significant stunt: it gets rid of file locking on index entry writes. This means that two different `fs.appendFile` calls could be happening to the exact same file at the same time, even in parallel on different processes.

My hypothesis is that this is totally fine, and the invariants it prevents are things cacache never actually promised. That is, concurrent writes will simply invalidate each other, and cacache does not promise the cache will always contain all data: it only guarantees that *if* the data is there, it is the *correct* data. As such, it's totally fine for some entries to corrupt and invalidate each other. Not to mention that the "expensive" operations involve the actual content, which does not invalidate so easily.

I'd still like some eyeballs on this. It still needs some tests. There's also some questions about this:

* are my assumptions correct that the only effect is mutual invalidation? Is there any case where two inserts can sync up perfectly to create one full valid entry with bizarre frankenstein data? What about 3 inserts? n inserts?
* by reducing bucket conflicts, the index can now have as many (or more) files in it than content/. Is this too many files? How many of them would there really be if the entire 4million+ entry npm registry were inserted into the cache?
* is a length check sufficient? Should it be a checksum instead?
* am I missing something here?
* Let's put it into npm, run a big install, and see how it goes, even if we have cacache-side tests.

Who needs lockfiles anyway? 😏 

### Benchmarks

Operation | Before | After | Change
---|---|---|---
`put()` | 1.686ms | 1.548ms | -8%
`put.stream()` | 1.364ms | 1.229ms | **-10%**
`index.insert()` | 0.5058ms | 0.3291ms | **-34%**
`index.find() hit` | 0.2518ms | 0.2603ms | *+3%*
`index.find() miss` | 0.1102ms | 0.1074ms | *-2.5%*